+++ aksel.nav.no/website/pages/god-praksis/artikler/[slug].tsx
@@ -1,12 +1,12 @@
-import NextLink from "next/link";
+import { differenceInMonths } from "date-fns";
 import { GetServerSideProps } from "next/types";
 import { Suspense, lazy } from "react";
-import { ChevronRightIcon } from "@navikt/aksel-icons";
 import { BodyLong, BodyShort, Detail, Heading, Label } from "@navikt/ds-react";
 import { UserStateT } from "@/auth/auth.types";
 import { getAuthUserState } from "@/auth/getUserState";
 import ArtikkelCard from "@/cms/cards/ArtikkelCard";
 import Footer from "@/layout/footer/Footer";
+import { GpTemaLink } from "@/layout/god-praksis-page/chipnavigation/GpTemaLink";
 import Header from "@/layout/header/Header";
 import { SanityBlockContent } from "@/sanity-block";
 import { getClient } from "@/sanity/client.server";
@@ -21,26 +21,30 @@ import {
   ResolveContributorsT,
   ResolveRelatedArticlesT,
   ResolveSlugT,
-  ResolveTemaT,
   TableOfContentsT,
 } from "@/types";
 import { abbrName, dateStr, generateTableOfContents } from "@/utils";
 import { Feedback } from "@/web/Feedback/Feedback";
+import OutdatedAlert from "@/web/OutdatedAlert";
 import { SEO } from "@/web/seo/SEO";
 import TableOfContents from "@/web/toc/TableOfContents";
-import NotFotfund from "../../404";
+import NotFound from "../../404";
 
 type PageProps = NextPageT<{
   page: ResolveContributorsT<
-    ResolveTemaT<ResolveSlugT<ResolveRelatedArticlesT<AkselGodPraksisDocT>>>
-  >;
+    ResolveSlugT<ResolveRelatedArticlesT<AkselGodPraksisDocT>>
+  > & {
+    innholdstype: string;
+    undertema: { title: string; tema: { slug: string; title: string } }[];
+  };
   publishDate: string;
   verifiedDate: string;
+  outdated: boolean;
   toc: TableOfContentsT;
   userState: UserStateT;
 }>;
 
-export const query = `{
+const query = `{
   "page": *[slug.current == $slug] | order(_updatedAt desc)[0]
   {
     ...,
@@ -49,7 +53,14 @@ export const query = `{
       ...,
       ${destructureBlocks}
     },
-    tema[]->{title, slug, seo},
+    "innholdstype": innholdstype->title,
+    "undertema": undertema[]->{
+      title,
+      "tema": tema->{
+        title,
+        "slug": slug.current,
+      }
+    },
     ${contributorsAll},
     relevante_artikler[]->{
       _id,
@@ -79,6 +90,9 @@ export const getServerSideProps: GetServerSideProps = async (
     slug: `god-praksis/artikler/${slug}`,
   });
 
+  const verifiedDate =
+    page?.updateInfo?.lastVerified ?? page?.publishedAt ?? page?._updatedAt;
+
   return {
     props: {
       page,
@@ -86,9 +100,8 @@ export const getServerSideProps: GetServerSideProps = async (
       preview: isPreview,
       id: page?._id ?? "",
       title: page?.heading ?? "",
-      verifiedDate: await dateStr(
-        page?.updateInfo?.lastVerified ?? page?.publishedAt ?? page?._updatedAt,
-      ),
+      verifiedDate: await dateStr(verifiedDate),
+      outdated: differenceInMonths(new Date(), new Date(verifiedDate)) >= 12,
       publishDate: await dateStr(page?.publishedAt ?? page?._updatedAt),
       toc: generateTableOfContents({
         content: page?.content,
@@ -104,11 +117,12 @@ const Page = ({
   page: data,
   publishDate,
   verifiedDate,
+  outdated,
   toc,
   userState,
 }: PageProps["props"]) => {
   if (!data) {
-    return <NotFotfund />;
+    return <NotFound />;
   }
 
   if (!data.content || !data.heading) {
@@ -121,8 +135,6 @@ const Page = ({
   }
 
   const authors = (data?.contributors as any)?.map((x) => x?.title) ?? [];
-
-  const hasTema = "tema" in data && data.tema && data?.tema.length > 0;
 
   const aside = data?.relevante_artikler &&
     data.relevante_artikler.length > 0 && (
@@ -154,9 +166,6 @@ const Page = ({
       </aside>
     );
 
-  const filteredTema =
-    hasTema && data?.tema?.filter((x: any) => x?.title && x?.slug);
-
   return (
     <>
       <SEO
@@ -176,10 +185,18 @@ const Page = ({
         <div className="mx-auto max-w-aksel px-4 sm:w-[90%]">
           <article className="pb-16 pt-12 md:pb-32">
             <div className="mx-auto mb-16 max-w-prose lg:ml-0">
+              {data.innholdstype && (
+                <BodyShort
+                  weight="semibold"
+                  className="uppercase text-violet-600"
+                >
+                  {data.innholdstype}
+                </BodyShort>
+              )}
               <Heading
                 level="1"
                 size="large"
-                className="text-wrap-balance mt-4 text-deepblue-800 md:text-5xl"
+                className="text-wrap-balance mt-1 text-deepblue-800 md:text-5xl"
               >
                 {data.heading}
               </Heading>
@@ -209,21 +226,17 @@ const Page = ({
                   </>
                 )}
               </div>
-              {hasTema && filteredTema && (
-                <div className="mt-8 flex flex-wrap gap-2">
-                  {filteredTema.map(({ title, slug }: any) => (
-                    <span key={title}>
-                      <BodyShort
-                        href={`/god-praksis/${slug.current}`}
-                        key={title}
-                        size="small"
-                        as={NextLink}
-                        className="flex min-h-8 items-center justify-center gap-[2px] rounded-full bg-surface-neutral-subtle pl-4  pr-1 text-deepblue-800 ring-1 ring-inset ring-border-subtle hover:bg-surface-neutral-subtle-hover focus:outline-none focus-visible:shadow-focus"
-                      >
-                        {title}
-                        <ChevronRightIcon aria-hidden fontSize="1.25rem" />
-                      </BodyShort>
-                    </span>
+              {data.undertema && data.undertema.length > 0 && (
+                <div className="mt-5 flex flex-wrap gap-2">
+                  {data.undertema.map(({ title, tema }) => (
+                    <GpTemaLink
+                      key={title}
+                      href={`/god-praksis/${
+                        tema.slug
+                      }?undertema=${encodeURIComponent(title)}`}
+                    >
+                      {tema.title}
+                    </GpTemaLink>
                   ))}
                 </div>
               )}
@@ -231,6 +244,7 @@ const Page = ({
             <div className="relative mx-auto mt-4 max-w-prose lg:ml-0 lg:grid lg:max-w-none lg:grid-flow-row-dense lg:grid-cols-3 lg:items-start lg:gap-x-12">
               <TableOfContents toc={toc} variant="subtle" />
               <div className="max-w-prose lg:col-span-2 lg:col-start-1">
+                {outdated && <OutdatedAlert />}
                 <SanityBlockContent blocks={data?.content ?? []} />
                 <div className="mt-12">
                   {authors?.length > 0 && (
